Update cpp tests
